+++ classifiers-options/src/main/java/com/ecaservice/classifier/options/model/NeuralNetworkOptions.java
@@ -12,6 +12,8 @@ import javax.validation.constraints.Min;
 import javax.validation.constraints.Pattern;
 import javax.validation.constraints.Size;
 
+import java.math.BigDecimal;
+
 import static com.ecaservice.classifier.options.model.Constraints.DECIMAL_VALUE_0_STRING;
 import static com.ecaservice.classifier.options.model.Constraints.DECIMAL_VALUE_1_STRING;
 import static com.ecaservice.classifier.options.model.Constraints.HIDDEN_LAYER_REGEX;
@@ -75,7 +77,7 @@ public class NeuralNetworkOptions extends ClassifierOptions {
     @DecimalMin(DECIMAL_VALUE_0_STRING)
     @DecimalMax(DECIMAL_VALUE_1_STRING)
     @Schema(description = "Min. error for optimization algorithm")
-    private Double minError;
+    private BigDecimal minError;
 
     /**
      * Activation function options
